pyproject.toml
[tool.poetry]
name = "masai"
version = "0.1.0"
description = "Multi-Agent Software AI"
authors = ["you"]

[tool.poetry.dependencies]
python = "^3.11"
groq = "^0.12.0"
instructor = "^1.3.0"
pydantic = "^2.0"
chromadb = "^0.5.0"
typer = "^0.12.0"
rich = "^13.0"
python-dotenv = "^1.0"

[tool.poetry.scripts]
masai = "main:app"
